bed bug fix

+++ api/Domain/Bed.cs
@@ -10,8 +10,8 @@ namespace API.Model
         [Key]
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public int Width { get; set; }
-        public int Length { get; set; }
+        public decimal Width { get; set; }
+        public decimal Length { get; set; }
         public virtual List<Cell> Cells { get; set; }
         public int NumOfColumns {  get; set; }
         public int NumOfRows { get; set; }
